Move Conway polys to their own module
