+++ src/galois/_polys/__init__.py
@@ -1,6 +1,7 @@
 """
 A subpackage containing arrays over Galois fields.
 """
+from ._conway import *
 from ._factor import *
 from ._functions import *
 from ._irreducible import *
